Fix AC button and decimal
The last commit introduce a bug with the AC button because the
switch case "clearAll" had been removed. If clearAll comes, the
switch will not receive it and treat it as a number.

The buttons changed colors.

Instead of rounding to 2 decimals, the calculator now includes up
to 7 numbers after the decimal point.